Moved Winner.component and implemented winner name in header
