+++ src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import {GameSetupComponent} from './components/pages/game-setup/game-setup.component';
 import {GameActiveComponent} from './components/pages/game-active/game-active.component';
-import {WinnerComponent} from './components/organisms/winner/winner.component';
+import {WinnerComponent} from './components/pages/winner/winner.component';
 
 const routes: Routes = [
     {
