--- src/app/components/organisms/winner/winner.component.ts
@@ -1,15 +1,0 @@
-import { Component } from '@angular/core';
-import {Router} from '@angular/router';
-
-@Component({
-  templateUrl: './winner.component.html'
-})
-export class WinnerComponent {
-
-  constructor(private readonly router: Router) {}
-
-  goBack() {
-    this.router.navigate(['game-setup']);
-  }
-
-}
