+++ src/app/components/pages/game-active/game-active.component.ts
@@ -52,7 +52,6 @@ export class GameActiveComponent implements AfterViewInit {
         this.thirdScore.points = '';
 
         if (value.finished) {
-          this.gameService.deleteGameById(this.gameId);
           this.router.navigate(['winner', this.gameId]);
         }
       });
