+++ src/app/components/pages/winner/winner.component.css
@@ -1,0 +1,5 @@
+h1 {
+  padding-top: 20px;
+  padding-left: 5%;
+}
+
