+++ src/app/components/pages/winner/winner.component.html
@@ -1,6 +1,4 @@
-<h1>Winner</h1>
+<h1>Winner: {{winner}}</h1>
 <div class="container">
   <app-button label="Nochmal spielen" (clicked)="goBack()"></app-button>
 </div>
-
-`
